create VisualStudioContextHelper.IsInServerModeAsync and add usage for credentials in VS
