+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/DefaultVSCredentialServiceProvider.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.Diagnostics;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
@@ -86,10 +87,13 @@ namespace NuGet.PackageManagement.VisualStudio
                 });
             }
 
+            // can only interact when VS is not in server mode.
+            bool nonInteractive = await VisualStudioContextHelper.IsInServerModeAsync(CancellationToken.None);
+
             // Initialize the credential service.
             var credentialService = new CredentialService(
                 new AsyncLazy<IEnumerable<ICredentialProvider>>(() => System.Threading.Tasks.Task.FromResult((IEnumerable<ICredentialProvider>)credentialProviders)),
-                nonInteractive: KnownUIContexts.CloudEnvironmentConnectedContext.IsActive, // cannot only interact when cloudconnection is not active.
+                nonInteractive: nonInteractive,
                 handlesDefaultCredentials: PreviewFeatureSettings.DefaultCredentialsAfterCredentialProviders);
 
             return credentialService;
